Streamlit/requirements.txt
streamlit
datetime
ibm_watsonx_ai==1.1.2
python-dotenv